# Conflicts:
#	ComGraphics/ComGraphics/Source/ItemStat.cpp

+++ ComGraphics/ComGraphics/Source/SceneEnd.cpp
@@ -398,7 +398,7 @@ void SceneEnd::RenderModelOnScreen(Mesh* mesh, float size, float Rotate, int rX,
 	modelStack.LoadIdentity(); //Reset modelStack
 	modelStack.Scale(size, size, size);
 	modelStack.Translate(x, y, z);
-	modelStack.Rotate(Rotate, rX, rY, rZ);
+	modelStack.Rotate(Rotate, (float)rX, (float)rY, (float)rZ);
 
 	RenderMesh(mesh, LightYN);
